requirements.txt
pdfkit==0.6.1
Flask==1.1.2
MarkupSafe==1.1.1
requests==2.23.0
python-twitter==3.5
gunicorn==20.0.4
psycopg2-binary==2.8.5
